+++ src/main/java/in/gov/abdm/nmr/db/sql/domain/college/ICollegeService.java
@@ -1,8 +1,9 @@
 package in.gov.abdm.nmr.db.sql.domain.college;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ICollegeService {
 
-    List<CollegeTO> getCollegeData(Long collegeId);
+    List<CollegeTO> getCollegeData(BigInteger collegeId);
 }
