+++ src/main/java/in/gov/abdm/nmr/db/sql/domain/sub_district/ISubDistrictService.java
@@ -1,8 +1,9 @@
 package in.gov.abdm.nmr.db.sql.domain.sub_district;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ISubDistrictService {
 
-    List<SubDistrictTO> getSubDistrictData(Long districtId);
+    List<SubDistrictTO> getSubDistrictData(BigInteger districtId);
 }
